requirements.txt
paho-mqtt
psycopg2-binary
redis
python-dotenv
